+++ packages/apptive_grid_form/lib/widgets/apptive_grid_form_widgets.dart
@@ -13,6 +13,7 @@ import 'package:apptive_grid_form/widgets/geolocation/geolocation_input.dart';
 import 'package:apptive_grid_form/widgets/geolocation/geolocation_map.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:http/http.dart' as http;
 import 'package:intl/intl.dart';
 import 'package:linked_scroll_controller/linked_scroll_controller.dart';
 import 'package:provider/provider.dart';